fix(test): address PR review findings
- Extract AioHTTPServer to shared helper.py with _thread.join()
- Add free_port() helper to avoid hardcoded port conflicts
- Add negative test: proxy with RUM disabled must NOT send
  x-datadog-rum-injection-pending header to upstream
- Assert injection-pending header doesn't leak into client response
- Assert server.stop() return value consistently
- Add assertion in test_http_proxy that injection-pending is absent
  in non-RUM builds
- Exclude requires_rum tests from docker-test target
