+++ test/integration-test/scenarios/conf/rum_proxy.conf
@@ -16,3 +16,8 @@ DatadogRum On
 </DatadogRumSettings>
 
 ProxyPass "/proxy" "${upstream_url}"
+
+<Location "/proxy-no-rum">
+  DatadogRum Off
+  ProxyPass "${upstream_url}"
+</Location>
